Completed assignment

+++ Binomial.java
@@ -22,7 +22,7 @@ return binomial1(n - 1, k) + binomial1(n - 1, k - 1);
 	
 	// Computes the Binomial function, efficiently
 	public static long binomial(int n, int k) {
-		long [][] memo = new int [n + 1][k + 1];
+		long [][] memo = new long [n + 1][k + 1];
 		 for(int i = 0; i < memo.length; i++) {
 			for(int j = 0; j < memo[0].length; j++) {
 				memo[i][j] = -1;
